--- backend/.gitignore
@@ -1,18 +1,0 @@
-# Node modules
-node_modules/
-
-# Environment variables
-.env
-
-# Logs
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# OS files
-.DS_Store
-Thumbs.db
-
-# Build output (if any)
-dist/
-build/
