--- backend/package.json
@@ -1,26 +1,0 @@
-{
-	"name": "backend",
-	"version": "1.0.0",
-	"description": "",
-	"main": "index.js",
-	"scripts": {
-		"dev": "nodemon server.js",
-		"start": "node server.js",
-		"test": "echo \"Error: no test specified\" && exit 1"
-	},
-	"keywords": [],
-	"author": "",
-	"license": "ISC",
-	"type": "commonjs",
-	"dependencies": {
-		"airtable": "^0.12.2",
-		"body-parser": "^2.2.0",
-		"cors": "^2.8.5",
-		"dotenv": "^17.2.2",
-		"express": "^5.1.0",
-		"stripe": "^18.5.0"
-	},
-	"devDependencies": {
-		"nodemon": "^3.1.10"
-	}
-}
