--- frontend/.gitignore
@@ -1,30 +1,0 @@
-# Node dependencies
-node_modules/
-
-# Build output
-build/
-dist/
-
-# Local environment variables
-.env
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-# Log files
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# Editor directories and files
-.vscode/
-.idea/
-*.sublime-workspace
-*.sublime-project
-
-# MacOS files
-.DS_Store
-
-# Misc
-*.log
