--- frontend/src/components/adModal/Modal.css
@@ -1,54 +1,0 @@
-/*  AD MODAL  */
-.ad-modal-overlay {
-    position: fixed;
-    z-index: 1000;
-    inset: 0;
-    background: rgba(68, 7, 71, 0.18);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.ad-modal {
-    background: #fff6fa;
-    border: 2px solid #EC83A8;
-    border-radius: 12px;
-    padding: 32px 24px 24px 24px;
-    max-width: 600px;
-    min-width: 500px;
-    box-shadow: 0 6px 32px #EC83A822;
-    position: relative;
-    color: #440747;
-    text-align: center;
-    animation: adModalPop 0.18s cubic-bezier(.4, 1.6, .6, 1) 1;
-}
-
-@keyframes adModalPop {
-    0% {
-        transform: scale(0.92);
-        opacity: 0;
-    }
-
-    100% {
-        transform: scale(1);
-        opacity: 1;
-    }
-}
-
-.ad-modal-close {
-    position: absolute;
-    top: 10px;
-    right: 14px;
-    background: none;
-    border: none;
-    font-size: 28px;
-    color: #EC83A8;
-    cursor: pointer;
-    line-height: 1;
-    padding: 0;
-    transition: color 0.2s;
-}
-
-.ad-modal-close:hover {
-    color: #BE1884;
-}
