--- frontend/src/components/adModal/benefitCard/BenefitCard.css
@@ -1,59 +1,0 @@
-.benefit-card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    /* center everything horizontally */
-    background: #ffe4ec;
-    border: 2px solid #ec83a8;
-    border-radius: 10px;
-    padding: 12px 18px;
-    margin: 18px 0 10px 0;
-    width: 100%;
-    /* fill parent container */
-    font-weight: 500;
-    font-size: 14px;
-    /* smaller font */
-    box-sizing: border-box;
-}
-
-.benefit-card-title {
-    width: 100%;
-    text-align: center;
-    /* center the title text */
-    margin-bottom: 10px;
-}
-
-.benefit-card-list {
-    display: flex;
-    flex-wrap: wrap;
-    /* allow items to wrap */
-    justify-content: center;
-    /* center items horizontally */
-    gap: 16px;
-    width: 100%;
-}
-
-.benefit-item {
-    display: flex;
-    align-items: center;
-    gap: 6px;
-    font-size: 13px;
-    /* shrink font */
-}
-
-.benefit-dot {
-    width: 18px;
-    height: 18px;
-    background: #fff;
-    border-radius: 50%;
-    border: 2px solid #be1884;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.benefit-check {
-    color: #be1884;
-    font-weight: bold;
-    font-size: 12px;
-}
