--- frontend/src/components/showcaseCard/ShowcaseCard.css
@@ -1,189 +1,0 @@
-/* SHOW CASE AD */
-.showcase-wrapper {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    align-items: center;
-    /* background: #fff6fa; */
-    background-color: #fff0f5;
-
-    /* pick a size that works for your layout */
-
-    transition: height 0.3s ease;
-    /* height: 100px; */
-    /* cap on mobile */
-    overflow: hidden;
-    /* keep hover effects inside */
-}
-
-/* SHOWCASE TOP ROW */
-.showcase-top-row {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    gap: 10px;
-
-}
-
-.showcase-title {
-    color: #BE1884;
-
-}
-
-.showcase-button {
-    font-weight: bold;
-    padding: 1px 5px;
-    border: 2px solid #EC83A8;
-    border-radius: 3px;
-    /* Means it takes up only space of title */
-    display: inline-block;
-    font-size: 12px;
-    box-sizing: content-box;
-    border-radius: 6px;
-    transition-duration: 0.2s;
-
-}
-
-.showcase-button:hover {
-    background-color: #EC83A8;
-    color: white;
-    cursor: pointer;
-}
-
-.showcase-view-button {
-    font-weight: bold;
-    padding: 1px 5px;
-    border: 2px solid #EC83A8;
-    border-radius: 3px;
-    /* Means it takes up only space of title */
-    display: inline-block;
-    font-size: 12px;
-    box-sizing: content-box;
-    border-radius: 6px;
-    margin-top: 10px;
-    transition-duration: 0.2s;
-
-}
-
-.showcase-view-button:hover {
-    background-color: #EC83A8;
-    color: white;
-    cursor: pointer;
-}
-
-
-/* SHOWCASE BOTTOM ROW */
-.showcase-ad-message {
-    margin-bottom: 5px;
-
-}
-
-.showcase-middle-row {
-    margin-top: 15px;
-    margin-bottom: 10px;
-}
-
-.clamp {
-    padding: 0 50px;
-    line-height: 1.3;
-    display: -webkit-box;
-    /* needed for line clamp */
-    -webkit-box-orient: vertical;
-    -webkit-line-clamp: 1;
-    line-clamp: 1;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    min-height: calc(1.3em * 1);
-    /* ensure 1 lines worth of height */
-    white-space: normal;
-    /* optional: for proper wrapping */
-    word-break: break-word;
-    /* optional: break long words */
-}
-
-.showcase-ad-link {
-    color: #EC83A8;
-    font-weight: bold;
-    text-decoration: none;
-    font-size: 15px;
-    transition: color 0.2s;
-}
-
-.showcase-ad-link-anonymous {
-    color: #dca8c8;
-    font-style: italic;
-    text-decoration: none;
-    font-size: 15px;
-}
-
-.showcase-ad-link:hover {
-    color: #BE1884;
-    text-decoration: underline;
-}
-
-/* SHOWCASE ARROW */
-.showcase-ad-arrow {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    width: 44px;
-    height: 100%;
-    border: none;
-    color: #BE1884;
-    font-size: 20px;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    transition: color 0.2s, background-color 0.2s, opacity 0.2s;
-    opacity: 0.7;
-    color: #EC83A8;
-    background-color: rgba(236, 131, 168, 0.15);
-    transition: background-color 0.3s;
-}
-
-.showcase-ad-arrow.left {
-    left: 0;
-}
-
-.showcase-ad-arrow.right {
-    right: 0;
-}
-
-/* Hover state: show background and full opacity */
-.showcase-ad-arrow:hover {
-    color: #EC83A8;
-    opacity: 1;
-    background-color: #ffe0f0;
-    /* hover color */
-    color: #BE1884;
-    transition: background-color 0.3s;
-}
-
-
-
-/* SHOWCASE DOTS */
-.showcase-ad-dots {
-    display: flex;
-    justify-content: center;
-    gap: 5px;
-    margin-top: 20px;
-}
-
-.showcase-ad-dot {
-    width: 7px;
-    height: 7px;
-    border-radius: 50%;
-    border: none;
-    background: #EC83A8;
-    opacity: 0.35;
-    cursor: pointer;
-    transition: opacity 0.2s, background 0.2s;
-}
-
-.showcase-ad-dot.active {
-    opacity: 1;
-    background: #BE1884;
-}
